python_tests/pyproject.toml
[project]
name = "osm-renderer-tests"
version = "0.1.0"
description = "Python regression test suite for Rust OSM renderer"
requires-python = ">=3.8"
dependencies = [
    "pytest>=8.0.0",
    "pytest-timeout>=2.2.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "osmium>=3.7.0",
    "psutil>=5.9.0",
    "pytest-html>=4.0.0",
    "pytest-xdist>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils", "test_data"]
